fix: SeriesRepository 수정

+++ backend/src/main/java/com/example/Flicktionary/domain/series/repository/SeriesRepository.java
@@ -3,5 +3,5 @@ package com.example.Flicktionary.domain.series.repository;
 import com.example.Flicktionary.domain.series.domain.Series;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SeriesRepository extends JpaRepository<Series, Integer> {
+public interface SeriesRepository extends JpaRepository<Series, Long> {
 }
